# BilfenAppPhp.RestAPI/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bilfen API",
        "version": "v1"
    },
    "components": {
        "schemas": {
...
            "Country": {
                "type": "object",
                "required": ["Id", "Name", "createdAt", "updatedAt"],
                "properties": {
                    "Id": { "type": "string", "description": "The id of the country" },
                    "Name": { "type": "string", "description": "The country's name" },
                    "createdAt": { "type": "string", "description": "The country's created date" },
                    "updatedAt": { "type": "string", "description": "The country's updated date" }
                },
                "example": {
                    "Id": "string",
                    "Name": "string",
                    "createdAt": "string",
                    "updatedAt": "string"
                }
            },
            "City": {
                "type": "object",
                "required": ["Id", "Name", "CountryId", "createdAt", "updatedAt"],
                "properties": {
                    "Id": { "type": "string", "description": "The id of the city" },
                    "Name": { "type": "string", "description": "The city's name" },
                    "createdAt": { "type": "string", "description": "The city's created date" },
                    "updatedAt": { "type": "string", "description": "The city's updated date" },
                    "CountryId": { "type": "string", "description": "The city's country id" },
                    "Country": { "$ref": "#/components/schemas/Country" }
                },
                "example": {
                    "Id": "string",
                    "Name": "string",
                    "createdAt": "string",
                    "updatedAt": "string",
                    "CountryId": "string",
                    "Country": {
                        "Id": "string",
                        "Name": "string",
                        "createdAt": "string",
                        "updatedAt": "string"
                    }
                }
            },
...
            "CreateUpdateCity": {
                "type": "object",
                "required": ["Id", "Name", "CountryId"],
                "properties": {
                    "Id": { "type": "string", "description": "The id of the city" },
                    "Name": { "type": "string", "description": "The city's name" },
                    "CountryId": { "type": "string", "description": "The city's country id" }
                }
            },
...
        }
    },
    "paths": {
...
        "/api/cities/GetAllCitiesByCountryId/{countryId}": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "parameters": [{
                    "in": "path",
                    "name": "countryId",
                    "description": "Country Id",
                    "schema": {
                        "type": "string"
                    },
                    "required": true
                }],
                "operationId": "allCitiesByCountryId",
                "responses": {
                    "200": {
                        "description": "Retrives a list of cities by country id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "$ref": "#/components/schemas/City" }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cities/GetCity/{id}": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "parameters": [{
                    "in": "path",
                    "name": "id",
                    "description": "City Id",
                    "schema": {
                        "type": "string"
                    },
                    "required": true
                }],
                "operationId": "getCity",
                "responses": {
                    "200": {
                        "description": "Retrives a city by id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/City"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cities/CreateCity": {
            "post": {
                "tags": ["Cities"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateCity"
                            }
                        }
                    }
                },
                "operationId": "createCity",
                "responses": {
                    "200": {
                        "description": "Create a new country data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/City"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cities/EditCity": {
            "put": {
                "tags": ["Cities"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateCity"
                            }
                        }
                    }
                },
                "operationId": "editCity",
                "responses": {
                    "200": {
                        "description": "Edit an existing city data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/City"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cities/DeleteCity/{id}": {
            "delete": {
                "tags": ["Cities"],
                "parameters": [{
                    "in": "path",
                    "name": "id",
                    "description": "City Id",
                    "schema": {
                        "type": "string"
                    },
                    "required": true
                }],
                "responses": {
                    "204": {
                        "description": "Delete an existing city data"
                    }
                }
            }
        },
...
    },
...
}